refactor: Rename Datadog filter to shouldKeepEvent
The extracted filter read awkwardly: discardExpectedInstanceErrors returned
true to *keep* an event and false to discard it, matching Datadog's
beforeSend contract (false discards) but contradicting its name.

Rename it to shouldKeepEvent — true keeps, false discards — so the name,
the return values, and the beforeSend contract all agree and it wires
straight in without a negation. Behavior is unchanged; the module and test
file are renamed to suit.
