--- src/integrations/datadog/discardExpectedInstanceErrors.test.ts
@@ -1,72 +1,0 @@
-import { describe, expect, it } from 'vitest';
-import { type DatadogErrorEvent, discardExpectedInstanceErrors } from './discardExpectedInstanceErrors';
-
-function errorEvent(error: DatadogErrorEvent['error']): DatadogErrorEvent {
-	return { type: 'error', error };
-}
-
-describe('discardExpectedInstanceErrors', () => {
-	it('keeps non-error events untouched', () => {
-		expect(discardExpectedInstanceErrors({ type: 'view' })).toBe(true);
-		expect(discardExpectedInstanceErrors({ type: 'resource' })).toBe(true);
-	});
-
-	it('keeps genuine application errors', () => {
-		expect(discardExpectedInstanceErrors(errorEvent({ message: 'TypeError: x is not a function' }))).toBe(true);
-	});
-
-	it('drops aborted requests regardless of endpoint', () => {
-		expect(discardExpectedInstanceErrors(errorEvent({ message: 'AxiosError: Request aborted' }))).toBe(false);
-	});
-
-	it('drops stray Fabric Connect breadcrumbs', () => {
-		expect(discardExpectedInstanceErrors(errorEvent({ message: 'Fabric Connect not established' }))).toBe(false);
-	});
-
-	// Regression test for issue #1371: handled AxiosError timeouts reach RUM via
-	// console.error with no resource URL, so they must be dropped on the message alone.
-	it('drops timeout errors even when no resource URL is present', () => {
-		expect(discardExpectedInstanceErrors(errorEvent({ message: 'AxiosError: timeout of 60000ms exceeded' }))).toBe(
-			false,
-		);
-		expect(discardExpectedInstanceErrors(errorEvent({ message: 'timeout of 15000ms exceeded' }))).toBe(false);
-	});
-
-	it('drops timeout errors against instance/cluster resource URLs', () => {
-		expect(
-			discardExpectedInstanceErrors(
-				errorEvent({
-					message: 'timeout of 60000ms exceeded',
-					resource: { url: 'https://api.harper.fast/HDBInstance/ins-123/operation' },
-				}),
-			),
-		).toBe(false);
-	});
-
-	it('drops network failures against an instance/cluster operation endpoint', () => {
-		expect(
-			discardExpectedInstanceErrors(
-				errorEvent({
-					message: 'Network Error',
-					resource: { url: 'https://api.harper.fast/Cluster/clu-123/operation' },
-				}),
-			),
-		).toBe(false);
-		expect(
-			discardExpectedInstanceErrors(
-				errorEvent({ source: 'network', resource: { url: 'https://api.harper.fast/HDBInstance/ins-9/operation' } }),
-			),
-		).toBe(false);
-	});
-
-	it('keeps non-timeout network failures that are not attributable to an instance endpoint', () => {
-		// Without an instance/cluster URL we cannot tell a real backend failure from an
-		// expected one, so a bare "Network Error" stays visible (unlike timeouts).
-		expect(discardExpectedInstanceErrors(errorEvent({ message: 'Network Error' }))).toBe(true);
-		expect(
-			discardExpectedInstanceErrors(
-				errorEvent({ message: 'Network Error', resource: { url: 'https://api.harper.fast/Organization/org-1' } }),
-			),
-		).toBe(true);
-	});
-});
